Auto play video, refactor
