+++ app/src/main/AndroidManifest.xml
@@ -19,7 +19,7 @@
     android:supportsRtl="true"
     android:theme="@style/Theme.Leanback">
     <activity
-      android:name=".OpaActivity"
+      android:name=".opa.OpaActivity"
       android:banner="@drawable/app_icon_your_company"
       android:icon="@drawable/app_icon_your_company"
       android:label="@string/title_activity_opa"
@@ -31,9 +31,9 @@
         <category android:name="android.intent.category.LEANBACK_LAUNCHER" />
       </intent-filter>
     </activity>
-    <activity android:name=".OpaDetailsActivity" />
-    <activity android:name=".PlaybackOverlayActivity" />
-    <activity android:name=".BrowseErrorActivity" />
+    <activity android:name=".details.OpaDetailsActivity" />
+    <activity android:name=".playback.PlaybackOverlayActivity" />
+    <activity android:name=".error.BrowseErrorActivity" />
   </application>
 
 </manifest>
