--- app/src/main/java/de/culo/lucyo/OpaActivity.java
@@ -1,33 +1,0 @@
-/*
- * Copyright (C) 2014 The Android Open Source Project
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License. You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software distributed under the License
- * is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
- * or implied. See the License for the specific language governing permissions and limitations under
- * the License.
- */
-
-package de.culo.lucyo;
-
-import android.app.Activity;
-import android.os.Bundle;
-
-/*
- * MainActivity class that loads MainFragment
- */
-public class OpaActivity extends Activity {
-    /**
-     * Called when the activity is first created.
-     */
-
-    @Override
-    public void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_opa);
-    }
-}
