+++ app/src/main/java/de/culo/lucyo/error/ErrorFragment.java
@@ -1,0 +1,89 @@
+/*
+ *                                       Manonilfu
+ *                                   turodiMedicalDime
+ *                               nsion,cheèunagrancazzata
+ *                       ,ioparlodellalo           cura,Ren
+ *                    è,lalocura.La                  pazzia,
+ *                  checazzoRenè,la                   cervez
+ *                  a,latradizione,o                   merda
+ *                  ,comelachia mitu,m    aconunabell  aspru
+ *                  zzatadipazzia:ilpeg giorconservator ismo
+ *                  cheperòsitingedis  impatia,dicolore,dipa
+ *                 illet  te.Inunapar  ola:Platinette;perché
+ *                Platinette,haicapito ,ciassolved atuttiino
+ *               strimali,datuttelen   ostremalefatte...Sono
+ *              cattolico,masonogiovaneevitaleperché  midiv
+ *             erton          oleminchiatedelsab     atoser
+ *            a.Ève                      roono?C     ifasen
+ *           tirela                                 coscie
+ *          nzaapo                                 stoPla
+ *         tinett                                  e,ques
+ *         taèl'                      Ital        iadelf
+ *         utur                      o:unp aes   edimus
+ *         iche                      tte,mentre  fuori                         c'èlamort
+ *        e!Èqu                      estochede  vifar                        etu:Occhidel
+ *        cuore                     sì,maconle suepa                       pparde    lle,
+ *        conle                     suetirate  contr                     oladrog    a,con
+ *        trol'                    abortomac  onunas                   trana,c     olora
+ *        ta,lu                    ccicante   frociaggine.Smalizia   taealle     graco
+ *         meun                   acazzodi    lambada.ÈlalocuraRenè,èlacaz      zodil
+ *         ocur                   a.Sel'a     cchia   ppiha   ivinto.Man      onilfu
+ *         turo                  diMedica      lDi   mension,cheèunagr      ancazz
+ *         ata,i               oparl odell         alocura,Renè,laloc     ura.Lap
+ *          azzi             a,che  cazzoRe         nè,lacerveza,latrad   izione,o
+ *          merda            ,comelachiamit                     u,maconu    nabellasp
+ *           ruzza            tadipazzia:i              lpeg       giorco  nser vator
+ *           ismoch              eper                   òsit        ingedi  simpatia
+ *            ,dicol                                ore              ,dipa    ille
+ *             tte.Inun                            apar              ola:P     lati
+ *                nette;p                          erch              éPlatinette,ha
+ *      ica        pito,ciass                       olve           datuttiinostrim
+ *     ali,dat    uttelenostremalef                  att         e...Son    o
+ *     cattolico,maso nogiovaneevitaleperch           émid    iverton
+ *     olem inchiatedelsab    atosera.Èveroon o?Cifasentirelacoscie
+ *      nzaa  postoPlati         nette,quest aèl'Italiadelfuturo
+ *       :unp   aesedi         musichette,m entre fuoric'èlam
+ *        orte!Èques           tochedevifa  retu
+ *         :Occhid              elcuores   ì,ma
+ *           con                lesuep    appa
+ *                               rdelle  ,con
+ *                                lesuetirat
+ *                                  econtro
+ *                                    lad
+ */
+package de.culo.lucyo.error;
+
+import android.os.Bundle;
+import android.util.Log;
+import android.view.View;
+
+import de.culo.lucyo.R;
+
+/*
+ * This class demonstrates how to extend ErrorFragment
+ */
+public class ErrorFragment extends android.support.v17.leanback.app.ErrorFragment {
+    private static final String TAG = "ErrorFragment";
+    private static final boolean TRANSLUCENT = true;
+
+    @Override
+    public void onCreate(Bundle savedInstanceState) {
+        Log.d(TAG, "onCreate");
+        super.onCreate(savedInstanceState);
+        setTitle(getResources().getString(R.string.app_name));
+    }
+
+    void setErrorContent() {
+        setImageDrawable(getResources().getDrawable(R.drawable.lb_ic_sad_cloud));
+        setMessage(getResources().getString(R.string.error_fragment_message));
+        setDefaultBackground(TRANSLUCENT);
+
+        setButtonText(getResources().getString(R.string.dismiss_error));
+        setButtonClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View arg0) {
+                getFragmentManager().beginTransaction().remove(ErrorFragment.this).commit();
+            }
+        });
+    }
+}
