+++ app/src/main/java/de/culo/lucyo/model/CardPresenter.java
@@ -1,0 +1,136 @@
+/*
+ *                                       Manonilfu
+ *                                   turodiMedicalDime
+ *                               nsion,cheèunagrancazzata
+ *                       ,ioparlodellalo           cura,Ren
+ *                    è,lalocura.La                  pazzia,
+ *                  checazzoRenè,la                   cervez
+ *                  a,latradizione,o                   merda
+ *                  ,comelachia mitu,m    aconunabell  aspru
+ *                  zzatadipazzia:ilpeg giorconservator ismo
+ *                  cheperòsitingedis  impatia,dicolore,dipa
+ *                 illet  te.Inunapar  ola:Platinette;perché
+ *                Platinette,haicapito ,ciassolved atuttiino
+ *               strimali,datuttelen   ostremalefatte...Sono
+ *              cattolico,masonogiovaneevitaleperché  midiv
+ *             erton          oleminchiatedelsab     atoser
+ *            a.Ève                      roono?C     ifasen
+ *           tirela                                 coscie
+ *          nzaapo                                 stoPla
+ *         tinett                                  e,ques
+ *         taèl'                      Ital        iadelf
+ *         utur                      o:unp aes   edimus
+ *         iche                      tte,mentre  fuori                         c'èlamort
+ *        e!Èqu                      estochede  vifar                        etu:Occhidel
+ *        cuore                     sì,maconle suepa                       pparde    lle,
+ *        conle                     suetirate  contr                     oladrog    a,con
+ *        trol'                    abortomac  onunas                   trana,c     olora
+ *        ta,lu                    ccicante   frociaggine.Smalizia   taealle     graco
+ *         meun                   acazzodi    lambada.ÈlalocuraRenè,èlacaz      zodil
+ *         ocur                   a.Sel'a     cchia   ppiha   ivinto.Man      onilfu
+ *         turo                  diMedica      lDi   mension,cheèunagr      ancazz
+ *         ata,i               oparl odell         alocura,Renè,laloc     ura.Lap
+ *          azzi             a,che  cazzoRe         nè,lacerveza,latrad   izione,o
+ *          merda            ,comelachiamit                     u,maconu    nabellasp
+ *           ruzza            tadipazzia:i              lpeg       giorco  nser vator
+ *           ismoch              eper                   òsit        ingedi  simpatia
+ *            ,dicol                                ore              ,dipa    ille
+ *             tte.Inun                            apar              ola:P     lati
+ *                nette;p                          erch              éPlatinette,ha
+ *      ica        pito,ciass                       olve           datuttiinostrim
+ *     ali,dat    uttelenostremalef                  att         e...Son    o
+ *     cattolico,maso nogiovaneevitaleperch           émid    iverton
+ *     olem inchiatedelsab    atosera.Èveroon o?Cifasentirelacoscie
+ *      nzaa  postoPlati         nette,quest aèl'Italiadelfuturo
+ *       :unp   aesedi         musichette,m entre fuoric'èlam
+ *        orte!Èques           tochedevifa  retu
+ *         :Occhid              elcuores   ì,ma
+ *           con                lesuep    appa
+ *                               rdelle  ,con
+ *                                lesuetirat
+ *                                  econtro
+ *                                    lad
+ */
+
+package de.culo.lucyo.model;
+
+import android.graphics.drawable.Drawable;
+import android.support.v17.leanback.widget.ImageCardView;
+import android.support.v17.leanback.widget.Presenter;
+import android.util.Log;
+import android.view.ViewGroup;
+
+import com.bumptech.glide.Glide;
+
+import de.culo.lucyo.R;
+
+/*
+ * A CardPresenter is used to generate Views and bind Objects to them on demand.
+ * It contains an Image CardView
+ */
+public class CardPresenter extends Presenter {
+    private static final String TAG = "CardPresenter";
+
+    private static final int CARD_WIDTH = 313;
+    private static final int CARD_HEIGHT = 176;
+    private static int sSelectedBackgroundColor;
+    private static int sDefaultBackgroundColor;
+    private Drawable mDefaultCardImage;
+
+    private static void updateCardBackgroundColor(ImageCardView view, boolean selected) {
+        int color = selected ? sSelectedBackgroundColor : sDefaultBackgroundColor;
+        // Both background colors should be set because the view's background is temporarily visible
+        // during animations.
+        view.setBackgroundColor(color);
+        view.findViewById(R.id.info_field).setBackgroundColor(color);
+    }
+
+    @Override
+    public ViewHolder onCreateViewHolder(ViewGroup parent) {
+        Log.d(TAG, "onCreateViewHolder");
+
+        sDefaultBackgroundColor = parent.getResources().getColor(R.color.default_background);
+        sSelectedBackgroundColor = parent.getResources().getColor(R.color.selected_background);
+        mDefaultCardImage = parent.getResources().getDrawable(R.drawable.movie);
+
+        ImageCardView cardView = new ImageCardView(parent.getContext()) {
+            @Override
+            public void setSelected(boolean selected) {
+                updateCardBackgroundColor(this, selected);
+                super.setSelected(selected);
+            }
+        };
+
+        cardView.setFocusable(true);
+        cardView.setFocusableInTouchMode(true);
+        updateCardBackgroundColor(cardView, false);
+        return new ViewHolder(cardView);
+    }
+
+    @Override
+    public void onBindViewHolder(Presenter.ViewHolder viewHolder, Object item) {
+        Movie movie = (Movie) item;
+        ImageCardView cardView = (ImageCardView) viewHolder.view;
+
+        Log.d(TAG, "onBindViewHolder");
+        if (movie.getCardImageUrl() != null) {
+            cardView.setTitleText(movie.getTitle());
+            cardView.setContentText(movie.getStudio());
+            cardView.setMainImageDimensions(CARD_WIDTH, CARD_HEIGHT);
+            Glide.with(viewHolder.view.getContext())
+                    .load(movie.getCardImageUrl())
+                    .centerCrop()
+                    .error(mDefaultCardImage)
+                    .into(cardView.getMainImageView());
+        }
+    }
+
+    @Override
+    public void onUnbindViewHolder(Presenter.ViewHolder viewHolder) {
+        Log.d(TAG, "onUnbindViewHolder");
+        ImageCardView cardView = (ImageCardView) viewHolder.view;
+        // Remove references to images so that the garbage collector can free up memory
+        cardView.setBadgeImage(null);
+        cardView.setMainImage(null);
+    }
+}
