+++ app/src/main/java/de/culo/lucyo/model/Movie.java
@@ -1,0 +1,183 @@
+/*
+ *                                       Manonilfu
+ *                                   turodiMedicalDime
+ *                               nsion,cheèunagrancazzata
+ *                       ,ioparlodellalo           cura,Ren
+ *                    è,lalocura.La                  pazzia,
+ *                  checazzoRenè,la                   cervez
+ *                  a,latradizione,o                   merda
+ *                  ,comelachia mitu,m    aconunabell  aspru
+ *                  zzatadipazzia:ilpeg giorconservator ismo
+ *                  cheperòsitingedis  impatia,dicolore,dipa
+ *                 illet  te.Inunapar  ola:Platinette;perché
+ *                Platinette,haicapito ,ciassolved atuttiino
+ *               strimali,datuttelen   ostremalefatte...Sono
+ *              cattolico,masonogiovaneevitaleperché  midiv
+ *             erton          oleminchiatedelsab     atoser
+ *            a.Ève                      roono?C     ifasen
+ *           tirela                                 coscie
+ *          nzaapo                                 stoPla
+ *         tinett                                  e,ques
+ *         taèl'                      Ital        iadelf
+ *         utur                      o:unp aes   edimus
+ *         iche                      tte,mentre  fuori                         c'èlamort
+ *        e!Èqu                      estochede  vifar                        etu:Occhidel
+ *        cuore                     sì,maconle suepa                       pparde    lle,
+ *        conle                     suetirate  contr                     oladrog    a,con
+ *        trol'                    abortomac  onunas                   trana,c     olora
+ *        ta,lu                    ccicante   frociaggine.Smalizia   taealle     graco
+ *         meun                   acazzodi    lambada.ÈlalocuraRenè,èlacaz      zodil
+ *         ocur                   a.Sel'a     cchia   ppiha   ivinto.Man      onilfu
+ *         turo                  diMedica      lDi   mension,cheèunagr      ancazz
+ *         ata,i               oparl odell         alocura,Renè,laloc     ura.Lap
+ *          azzi             a,che  cazzoRe         nè,lacerveza,latrad   izione,o
+ *          merda            ,comelachiamit                     u,maconu    nabellasp
+ *           ruzza            tadipazzia:i              lpeg       giorco  nser vator
+ *           ismoch              eper                   òsit        ingedi  simpatia
+ *            ,dicol                                ore              ,dipa    ille
+ *             tte.Inun                            apar              ola:P     lati
+ *                nette;p                          erch              éPlatinette,ha
+ *      ica        pito,ciass                       olve           datuttiinostrim
+ *     ali,dat    uttelenostremalef                  att         e...Son    o
+ *     cattolico,maso nogiovaneevitaleperch           émid    iverton
+ *     olem inchiatedelsab    atosera.Èveroon o?Cifasentirelacoscie
+ *      nzaa  postoPlati         nette,quest aèl'Italiadelfuturo
+ *       :unp   aesedi         musichette,m entre fuoric'èlam
+ *        orte!Èques           tochedevifa  retu
+ *         :Occhid              elcuores   ì,ma
+ *           con                lesuep    appa
+ *                               rdelle  ,con
+ *                                lesuetirat
+ *                                  econtro
+ *                                    lad
+ */
+
+package de.culo.lucyo.model;
+
+import android.util.Log;
+
+import java.io.Serializable;
+import java.net.URI;
+import java.net.URISyntaxException;
+
+/*
+ * Movie class represents video entity with title, description, image thumbs and video url.
+ *
+ */
+public class Movie implements Serializable {
+    static final long serialVersionUID = 727566175075960653L;
+    private static long count = 0;
+    private long id;
+    private String title;
+    private String description;
+    private String bgImageUrl;
+    private String cardImageUrl;
+    private String videoUrl;
+    private String studio;
+    private String category;
+
+    public Movie() {
+    }
+
+    public static long getCount() {
+        return count;
+    }
+
+    public static void incCount() {
+        count++;
+    }
+
+    public long getId() {
+        return id;
+    }
+
+    public void setId(long id) {
+        this.id = id;
+    }
+
+    public String getTitle() {
+        return title;
+    }
+
+    public void setTitle(String title) {
+        this.title = title;
+    }
+
+    public String getDescription() {
+        return description;
+    }
+
+    public void setDescription(String description) {
+        this.description = description;
+    }
+
+    public String getStudio() {
+        return studio;
+    }
+
+    public void setStudio(String studio) {
+        this.studio = studio;
+    }
+
+    public String getVideoUrl() {
+        return videoUrl;
+    }
+
+    public void setVideoUrl(String videoUrl) {
+        this.videoUrl = videoUrl;
+    }
+
+    public String getBackgroundImageUrl() {
+        return bgImageUrl;
+    }
+
+    public void setBackgroundImageUrl(String bgImageUrl) {
+        this.bgImageUrl = bgImageUrl;
+    }
+
+    public String getCardImageUrl() {
+        return cardImageUrl;
+    }
+
+    public void setCardImageUrl(String cardImageUrl) {
+        this.cardImageUrl = cardImageUrl;
+    }
+
+    public String getCategory() {
+        return category;
+    }
+
+    public void setCategory(String category) {
+        this.category = category;
+    }
+
+    public URI getBackgroundImageURI() {
+        try {
+            Log.d("BACK MOVIE: ", bgImageUrl);
+            return new URI(getBackgroundImageUrl());
+        } catch (URISyntaxException e) {
+            Log.d("URI exception: ", bgImageUrl);
+            return null;
+        }
+    }
+
+    public URI getCardImageURI() {
+        try {
+            return new URI(getCardImageUrl());
+        } catch (URISyntaxException e) {
+            return null;
+        }
+    }
+
+    @Override
+    public String toString() {
+        return "Movie{" +
+                "id=" + id +
+                ", title='" + title + '\'' +
+                ", videoUrl='" + videoUrl + '\'' +
+                ", backgroundImageUrl='" + bgImageUrl + '\'' +
+                ", backgroundImageURI='" + getBackgroundImageURI().toString() + '\'' +
+                ", cardImageUrl='" + cardImageUrl + '\'' +
+                '}';
+    }
+}
