+++ app/src/main/res/layout/activity_opa.xml
@@ -2,9 +2,9 @@
 <fragment xmlns:android="http://schemas.android.com/apk/res/android"
   xmlns:tools="http://schemas.android.com/tools"
   android:id="@+id/main_browse_fragment"
-  android:name="de.culo.lucyo.OpaFragment"
+  android:name="de.culo.lucyo.opa.OpaFragment"
   android:layout_width="match_parent"
   android:layout_height="match_parent"
-  tools:context=".OpaActivity"
+  tools:context=".opa.OpaActivity"
   tools:deviceIds="tv"
   tools:ignore="MergeRootFrame" />
