+++ app/src/main/res/layout/playback_controls.xml
@@ -26,7 +26,7 @@
 
   <fragment
     android:id="@+id/playback_controls_fragment"
-    android:name="de.culo.lucyo.PlaybackOverlayFragment"
+    android:name="de.culo.lucyo.playback.PlaybackOverlayFragment"
     android:layout_width="match_parent"
     android:layout_height="match_parent" />
